app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
uvicorn = "^0.30.4"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
sqlalchemy-utils = "^0.41.2"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"
faker = "^26.1.0"
pydantic-settings = "^2.4.0"
sentry-sdk = "^2.12.0"
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
pillow = "^10.4.0"
pytz = "^2024.1"
python-magic = "^0.4.27"
flake8 = "^7.1.0"
mypy = "^1.11.1"
flake8-quotes = "^3.4.0"
flake8-bandit = "^4.1.1"
flake8-broken-line = "^1.0.0"
flake8-bugbear = "^24.4.26"
flake8-commas = "^4.0.0"
flake8-comprehensions = "^3.15.0"
flake8-debugger = "^4.1.2"
flake8-docstrings = "^1.7.0"
flake8-eradicate = "^1.5.0"
flake8-in-file-ignores = "^0.2.2"
flake8-isort = "^6.1.1"
flake8-polyfill = "^1.0.2"
flake8-rst-docstrings = "^0.3.0"
flake8-string-format = "^0.3.0"
black = "^24.8.0"
yarl = "^1.9.4"
fastapi-pagination = "^0.12.26"
autoflake = "^2.3.1"
isort = "^5.13.2"
geoalchemy2 = "^0.15.2"
pytest-xdist = "^3.6.1"
saq = {extras = ["hiredis", "web"], version = "^0.13.3"}
types-pytz = "^2024.1.0.20240417"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
